Moved redis functionalities to a single file. Tested methods on ts-node REPL.
