+++ package.json
@@ -22,8 +22,6 @@
     "nodemon": "^3.1.4",
     "redis": "^4.7.0",
     "shoukaku": "^4.1.1",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.5.4",
     "winston": "^3.14.2",
     "winston-transport": "^4.7.1"
   },
@@ -31,6 +29,8 @@
     "@types/express": "^4.17.21",
     "@types/node": "^22.5.2",
     "prettier": "^3.3.3",
-    "tsconfig-paths": "^4.2.0"
+    "ts-node": "^10.9.2",
+    "tsconfig-paths": "^4.2.0",
+    "typescript": "^5.6.2"
   }
 }
